+++ app/src/main/java/com/ayoprez/castro/repository/EventsRepository.java
@@ -8,7 +8,7 @@ import java.util.ArrayList;
  * Created by ayo on 26.06.16.
  */
 public interface EventsRepository {
-    EventItem getEvent(short id);
+    EventItem getEvent(int id);
     ArrayList<EventItem> getAllEvents();
     int getEventIdByPosition(int position);
     EventItem getEventByPosition(int position);
